integer share price (now in tez)

+++ artcrowd/migrations/0001_initial.py
@@ -69,7 +69,7 @@ class Migration(migrations.Migration):
                 ('created_on', models.DateTimeField(auto_now_add=True)),
                 ('deadline', models.DateTimeField()),
                 ('status', models.CharField(choices=[('new', 'new'), ('approved by admin', 'approved by admin'), ('approved by artist', 'approved by artist'), ('rejected', 'rejected'), ('expired', 'expired'), ('closed', 'closed')], default='new', max_length=50)),
-                ('share_price', models.DecimalField(decimal_places=6, max_digits=16)),
+                ('share_price', models.IntegerField()),
                 ('min_shares', models.IntegerField(blank=True, null=True)),
                 ('max_shares', models.IntegerField(blank=True, null=True)),
                 ('royalty_pct', models.IntegerField(blank=True, default=0)),
